Changed pm interception point for rpc requests
Jira issue: WP-463
