+++ webinos/common/rpc/lib/rpc.js
@@ -22,14 +22,14 @@
 	if (typeof webinos === 'undefined')
 		webinos = {};
 	var logger = console;
-    var policyManager = {'enforceRPCRequest':function(scope, jsonRPC, from, msgid){handleRequest.call(scope, jsonRPC, from, msgid)}};
+//    var policyManager = {'enforceRPCRequest':function(scope, jsonRPC, from, msgid){handleRequest.call(scope, jsonRPC, from, msgid)}};
 	if (typeof module === 'undefined') {
 		var exports = {};
 	} else {
 		var exports = module.exports = {};
 		var webinos_= require("find-dependencies")(__dirname);
 		logger  = webinos_.global.require(webinos_.global.util.location, "lib/logging.js")(__filename);
-        policyManager = webinos_.global.require(webinos_.global.manager.policy_manager.location).policyManager;
+//        policyManager = webinos_.global.require(webinos_.global.manager.policy_manager.location).policyManager;
 	}	
 
 	var idCount = 0;
@@ -296,8 +296,8 @@
 
 		if (typeof jsonRPC.method !== 'undefined' && jsonRPC.method != null) {
 			// received message is RPC request
-            policyManager.enforceRPCRequest(this, jsonRPC, from, msgid, handleRequest);
-//			handleRequest.call(this, jsonRPC, from, msgid);
+//            policyManager.enforceRPCRequest(this, jsonRPC, from, msgid, handleRequest);
+			handleRequest.call(this, jsonRPC, from, msgid);
 		} else {
 			// received message is RPC response
 			handleResponse.call(this, jsonRPC, from, msgid);
